Move `calculateSSHLogins` to `lib/client`
